+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20240920/ApiModels/Alert.cs
@@ -1,4 +1,4 @@
-namespace TeachingRecordSystem.Api.V3.VNext.ApiModels;
+namespace TeachingRecordSystem.Api.V3.V20240920.ApiModels;
 
 [AutoMap(typeof(Core.SharedModels.Alert))]
 public record Alert