+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20240920/Requests/SetDeceasedRequest.cs
@@ -1,4 +1,4 @@
-namespace TeachingRecordSystem.Api.V3.VNext.Requests;
+namespace TeachingRecordSystem.Api.V3.V20240920.Requests;
 
 public class SetDeceasedRequest
 {