+++ TeachingRecordSystem/tests/TeachingRecordSystem.Api.Tests/V3/V20240920/FindPersonByLastNameAndDateOfBirthTests.cs
@@ -1,9 +1,9 @@
-namespace TeachingRecordSystem.Api.Tests.V3.VNext;
+namespace TeachingRecordSystem.Api.Tests.V3.V20240920;
 
 [Collection(nameof(DisableParallelization))]
-public class FindPersonsByTrnAndDateOfBirthTests : TestBase
+public class FindPersonByLastNameAndDateOfBirthTests : TestBase
 {
-    public FindPersonsByTrnAndDateOfBirthTests(HostFixture hostFixture) : base(hostFixture)
+    public FindPersonByLastNameAndDateOfBirthTests(HostFixture hostFixture) : base(hostFixture)
     {
         XrmFakedContext.DeleteAllEntities<Contact>();
         SetCurrentApiClient([ApiRoles.GetPerson]);
@@ -13,7 +13,6 @@ public class FindPersonsByTrnAndDateOfBirthTests : TestBase
     public async Task Get_ValidRequestWithMatchOnPersonWithAlerts_ReturnsExpectedAlertsContent()
     {
         // Arrange
-        var findBy = "LastNameAndDateOfBirth";
         var lastName = "Smith";
         var dateOfBirth = new DateOnly(1990, 1, 1);
 
@@ -30,9 +29,20 @@ public class FindPersonsByTrnAndDateOfBirthTests : TestBase
 
         var sanction = person.Sanctions.Single();
 
-        var request = new HttpRequestMessage(
-            HttpMethod.Get,
-            $"/v3/persons?findBy={findBy}&lastName={lastName}&dateOfBirth={dateOfBirth:yyyy-MM-dd}");
+        var request = new HttpRequestMessage(HttpMethod.Post, $"/v3/persons/find")
+        {
+            Content = JsonContent.Create(new
+            {
+                persons = new[]
+                {
+                    new
+                    {
+                        trn = person.Trn,
+                        dateOfBirth = person.DateOfBirth
+                    }
+                }
+            })
+        };
 
         // Act
         var response = await GetHttpClientWithApiKey().SendAsync(request);